requirements.txt
discord.py
flask
waitress